[Sukumar] Fixes Token repository not acting like a singleton. Fixes login button state for MapScreen flow from login and Bid to buy

+++ lib/models/item.dart
@@ -19,7 +19,7 @@ class Item {
       name: itemName,
       displayName: itemDisplayNames[itemName],
       qty: json['quantity'],
-      price: json['cost'].toDouble(),
+      price: json['cost'],
     );
   }
 
@@ -45,7 +45,7 @@ class Item {
   String name;
   String displayName;
   int qty;
-  double price;
+  int price;
 
   Map<String, dynamic> toJson() => {
         'qty': qty,
